+++ payments-api/bootstrap/app.go
@@ -13,6 +13,7 @@ import (
 
 	"github.com/jtonynet/go-payments-api/internal/adapter/database"
 	"github.com/jtonynet/go-payments-api/internal/adapter/inMemoryDatabase"
+	"github.com/jtonynet/go-payments-api/internal/adapter/pubSub"
 	"github.com/jtonynet/go-payments-api/internal/adapter/repository"
 
 	"github.com/jtonynet/go-payments-api/internal/core/port"
@@ -38,18 +39,13 @@ func NewApp(cfg *config.Config) (App, error) {
 	}
 	app.Logger = logger
 
-	cacheCfg, _ := cfg.Cache.ToInMemoryDatabase()
-	cacheConn, err := inMemoryDatabase.NewConn(cacheCfg)
+	pubSubUnlock, err := pubSub.New(cfg.PubSub)
 	if err != nil {
-		return App{}, fmt.Errorf("error: dont instantiate cache client: %v", err)
-	}
-
-	if cacheConn.Readiness(context.TODO()) != nil {
-		return App{}, fmt.Errorf("error: dont connecting to cache: %v", err)
+		return App{}, fmt.Errorf("error: dont instantiate pubsub client: %v", err)
 	}
 
 	lockCfg, _ := cfg.Lock.ToInMemoryDatabase()
-	lockConn, err := inMemoryDatabase.NewConn(lockCfg)
+	lockConn, err := inMemoryDatabase.NewClient(lockCfg)
 	if err != nil {
 		return App{}, fmt.Errorf("error: dont instantiate lock client: %v", err)
 	}
@@ -59,7 +55,17 @@ func NewApp(cfg *config.Config) (App, error) {
 	}
 
 	if logger != nil {
-		logger.Debug("successfully: connected to the cache!")
+		logger.Debug("successfully: connected to the lock!")
+	}
+
+	cacheCfg, _ := cfg.Cache.ToInMemoryDatabase()
+	cacheConn, err := inMemoryDatabase.NewClient(cacheCfg)
+	if err != nil {
+		return App{}, fmt.Errorf("error: dont instantiate cache client: %v", err)
+	}
+
+	if cacheConn.Readiness(context.TODO()) != nil {
+		return App{}, fmt.Errorf("error: dont connecting to cache: %v", err)
 	}
 
 	dbConn, err := database.NewConn(cfg.Database)
@@ -88,7 +94,7 @@ func NewApp(cfg *config.Config) (App, error) {
 		return App{}, fmt.Errorf("error: dont instantiate merchant cached repository: %v", err)
 	}
 
-	memoryLockRepo, err := repository.NewMemoryLock(lockConn)
+	memoryLockRepo, err := repository.NewMemoryLock(lockConn, pubSubUnlock)
 	if err != nil {
 		return App{}, fmt.Errorf("error: dont instantiate memory lock repository: %v", err)
 	}